+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.0.22
+
+* Add opportunity to customize table header cells
+
 ## 0.0.21
 
 * Upfill `UIKitWidth`